% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-aggregate_datasets.R
\name{aggregate_result_URLs}
\alias{aggregate_result_URLs}
\title{Aggregate BacDive-URLs from a Paged List of Retrieved URLs}
\usage{
aggregate_result_URLs(results)
}
\arguments{
\item{results}{A list of paginated URLs returned by an ambiguous
\code{bd_retrieve_…()} call.}
}
\value{
An integer vector of all BacDive URLs within the results.
}
\description{
Aggregate BacDive-URLs from a Paged List of Retrieved URLs
}
\keyword{internal}
