% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{authenticate}
\alias{authenticate}
\title{authenticate to BacDive}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{character(1) BacDive email}

\item{password}{character(1) BacDive password}
}
\value{
A BacDive access object
}
\description{
Credentials should be stored in environment variables "BACDIVE_USERNAME" and
"BACDIVE_PASSWORD".
}
\examples{
Sys.setenv("BACDIVE_USERNAME" = "my.user.name@my.school.edu",
           "BACDIVE_PASSWORD" = "my password")
ao <- authenticate(Sys.getenv("BACDIVE_USERNAME"),
                   Sys.getenv("BACDIVE_PASSWORD"))
}
