% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_retrieve.R, R/bd_retrieve_by_culture.R,
%   R/bd_retrieve_by_sequence.R, R/bd_retrieve_taxon.R
\name{bd_retrieve}
\alias{bd_retrieve}
\alias{bd_retrieve_by_culture}
\alias{bd_retrieve_by_sequence}
\alias{bd_retrieve_taxon}
\title{Retrieve (a) Dataset(s) from BacDive}
\usage{
bd_retrieve(id)

bd_retrieve_by_culture(collection_no)

bd_retrieve_by_sequence(accession)

bd_retrieve_taxon(name)
}
\arguments{
\item{id}{Mandatory character string or numeric of the BacDive strain ID
whose data you want to retrieve.}

\item{collection_no}{Mandatory character string of the culture collection
number.}

\item{accession}{Mandatory character string of a sequence accession number.}

\item{name}{Mandatory character string, specifying the taxon. Abbreviations
like "B. halotolerans" are not allowed!}
}
\value{
A (large) list containing either a single BacDive dataset in case
the search term was unambiguous, or if ambiguous: all matching datasets.
Read \href{https://bacdive.dsmz.de/api/bacdive/}{BacDive.DSMZ.de/api/bacdive}
to learn more.
}
\description{
Retrieve (a) Dataset(s) from BacDive
}
\examples{
\dontrun{
\donttest{
dataset_717 <- bd_retrieve(id = 717)
dataset_717 <- bd_retrieve(id = "717")
}}
\dontrun{
\donttest{
dataset_DSM_319 <- bd_retrieve_by_culture(collection_no = "DSM 319")
}}
\dontrun{
\donttest{
dataset_AJ000733 <- bd_retrieve_by_sequence(accession = "AJ000733")
}}
\dontrun{
\donttest{
datasets_Bh <- bd_retrieve_taxon(name = "Bacillus halotolerans")
datasets_Bss <- bd_retrieve_taxon(name = "Bacillus subtilis subtilis")
}}
}
