% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_search_results.R
\name{bd_retrieve_by_search}
\alias{bd_retrieve_by_search}
\title{Retrieve Results of an Advanced Search/Query}
\usage{
bd_retrieve_by_search(queryURL)
}
\arguments{
\item{queryURL}{URL of a query prepared on \href{https://bacdive.dsmz.de/advsearch}{BacDive.DSMZ.de/AdvSearch},
as explained in the vignette "Pre-Configuring Advanced Searches and Retrieving the Results".}
}
\description{
Retrieve Results of an Advanced Search/Query
}
\examples{
\dontrun{
\donttest{
plant_pathogens <- bd_retrieve_by_search(queryURL = "https://bacdive.dsmz.de/advsearch?advsearch=search&site=advsearch&searchparams\%5B5\%5D\%5Bsearchterm\%5D=1&searchparams\%5B156\%5D\%5Bsearchterm\%5D=&searchparams\%5B158\%5D\%5Bsearchterm\%5D=1")
}}
}
