% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-construct_Renviron_path.R
\name{construct_Renviron_path}
\alias{construct_Renviron_path}
\title{Construct Path of User's .Renviron File}
\usage{
construct_Renviron_path()
}
\value{
A character vector representing the path to the user's \code{.Renviron}
file.
}
\description{
Construct Path of User's .Renviron File
}
