% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-construct_url.R
\name{construct_url}
\alias{construct_url}
\title{Construct URLs of Web Service searches}
\usage{
construct_url(searchTerm, searchType = "bacdive_id")
}
\description{
Construct URLs of Web Service searches
}
\keyword{internal}
