% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.downloadCSV}
\alias{.downloadCSV}
\title{Download CSV with BacDive IDs}
\usage{
.downloadCSV(
  bacdive_url = "https://bacdive.dsmz.de/advsearch/csv",
  update_cache = FALSE
)
}
\arguments{
\item{bacdive_url}{character(1) URL to BacDive CSV. Defaults to
https://bacdive.dsmz.de/advsearch/csv.}

\item{update_cache}{bool Downloads the file again if TRUE. Defaults to FALSE.}
}
\description{
Download CSV with BacDive IDs
}
\examples{
downloadCSV()
downloadCSV("https://bacdive.dsmz.de/advsearch/csv")
}
