% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.formatRecord}
\alias{.formatRecord}
\title{Use template CSV to get elements or functions to process BacDive data}
\usage{
.formatRecord(bacdive_data, record_template, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a bacdive data}

\item{record_template}{data.frame of instructions for each row of data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
list of data
}
\description{
Use template CSV to get elements or functions to process BacDive data
}
\examples{
formatRecord(bacdive_data, template)
}
