% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getAccession}
\alias{.getAccession}
\title{Get the accession id}
\usage{
.getAccession(sequence_info, pattern, verbose = FALSE)
}
\arguments{
\item{sequence_info}{list of sequence data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Get the accession id
}
\examples{
.getAccession(sequence_info, TRUE)
}
