% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getAll}
\alias{.getAll}
\title{Get all data and transform into JSON}
\usage{
.getAll(bacdive_data, elements, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{elements}{vector of strings of elements in BacDive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Get all data and transform into JSON
}
\examples{
.getAll(bacdive_data, elements, TRUE)
}
