% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getClassification}
\alias{.getClassification}
\title{Use LPSN if available for classification}
\usage{
.getClassification(bacdive_data, elements, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{elements}{vector of strings of elements in BacDive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
list
}
\description{
Use LPSN if available for classification
}
\examples{
.getClassification(bacdive_data, elements, TRUE)
}
