% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getHemolysis}
\alias{.getHemolysis}
\title{Get hemolysis data}
\usage{
.getHemolysis(bacdive_data, elements, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{elements}{vector of strings of elements in BacDive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Data will be in the format: metabolite production
Data will be separated by semicolons.
}
\examples{
.getHemolysis(bacdive_data, elements)
}
