% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getMetabolites}
\alias{.getMetabolites}
\title{Get metabolite data, which are in lists of lists}
\usage{
.getMetabolites(bacdive_data, elements, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{elements}{vector of strings of elements in BacDive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Separate multiple values by a semicolon
}
\examples{
.getMetabolites(bacdive_data, elements, TRUE)
}
