% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getMetabolitesUtilization}
\alias{.getMetabolitesUtilization}
\title{Get metabolite utilization data}
\usage{
.getMetabolitesUtilization(metabolites_entries, verbose = verbose)
}
\arguments{
\item{verbose}{logical print messages. Default to FALSE.}

\item{metabolite_entries}{list of metabolites entries}
}
\value{
vector of character
}
\description{
Data will be in the format: metabolite utilization_activity
kind_of_utilization_tested
}
\examples{
.getMetabolitesUtilization(metabolites_entries, TRUE)
}
