% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getNcbiId}
\alias{.getNcbiId}
\title{Get NCBI ID}
\usage{
.getNcbiId(bacdive_data, level = "strain", verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{level}{character representing the order. Defaults to 'strain'.}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
If BacDive doesn't have NCBI ID in 'General', try searching by the taxon_name
in taxizedb.
}
\examples{
.getTaxonName(bacdive_data, TRUE)
}
