% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getParent}
\alias{.getParent}
\title{Get parent taxon name}
\usage{
.getParent(bacdive_data, elements, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{elements}{character representing the elements column of the template}

\item{verbose}{logical print messages. Default to FALSE.}

\item{database}{character representing the sequence databases in
.getSequenceDatabases().}
}
\value{
vector of character
}
\description{
If the species doesn't exist, check the next rank up in .getValidRank()
}
\examples{
.getParent(bacdive_data, "Name and taxonomic classification")
}
