% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getParentNcbiId}
\alias{.getParentNcbiId}
\title{Get parent NCBI id}
\usage{
.getParentNcbiId(bacdive_data, elements, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Get parent NCBI id
}
\examples{
.getParent(bacdive_data, "Name and taxonomic classification")
}
