% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getParentRank}
\alias{.getParentRank}
\title{Get parent rank}
\usage{
.getParentRank(bacdive_data, elements, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Get parent rank
}
\examples{
.getParentRank(bacdive_data, , "Name and taxonomic classification")
}
