% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getScientificName}
\alias{.getScientificName}
\title{Get the Scientific Name}
\usage{
.getScientificName(bacdive_data, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a bacdive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
a character vector
}
\description{
Try the LPSN (https://lpsn.dsmz.de/) path first. Otherwise, choose first
reference
}
