% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getSequenceId}
\alias{.getSequenceId}
\title{Get the genome or 16S sequence id}
\usage{
.getSequenceId(bacdive_data, sequence_type, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Get the genome or 16S sequence id
}
\examples{
.getAccession(bacdive_data, TRUE)
}
