% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getValue}
\alias{.getValue}
\title{Get value if it exists}
\usage{
.getValue(value, ..., verbose = FALSE)
}
\arguments{
\item{value}{A list of lists of data}

\item{...}{A list}

\item{verbose}{logical message if TRUE. Defaults to FALSE.}
}
\value{
integer string character(0)
}
\description{
Get value if it exists
}
\examples{
getValue(result, 'General', 'BacDive-ID')
}
