% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getValues}
\alias{.getValues}
\title{Get values if any exist}
\usage{
.getValues(value, ..., as_json = FALSE, verbose = FALSE)
}
\arguments{
\item{value}{A list of lists of data}

\item{...}{A list}

\item{as_json}{logical if TRUE, return as JSON. Defaults to FALSE.}

\item{verbose}{logical message if TRUE. Defaults to FALSE.}
}
\value{
integer string character(0) list
}
\description{
Get values if any exist
}
\examples{
getValues(result, 'General', 'BacDive-ID')
}
