% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.readCSV}
\alias{.readCSV}
\title{Read BacDive CSV and return a list of BacDive IDs}
\usage{
.readCSV(bacdive_csv = "BacDiveIDs.csv")
}
\arguments{
\item{bacdive_csv}{character(1) CSV with BacDive IDs. Defaults to
BacDiveIDs.csv.}
}
\value{
list of integers BacDive IDs
}
\description{
Read BacDive CSV and return a list of BacDive IDs
}
\examples{
readCSV(bacdive_csv)
}
