% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.readTemplate}
\alias{.readTemplate}
\title{Read Template CSV for creating rows from BacDive data}
\usage{
.readTemplate(path)
}
\arguments{
\item{path}{character(1) to template.csv}
}
\value{
data.frame of instructions per row
}
\description{
Read Template CSV for creating rows from BacDive data
}
\examples{
template <- readTemplate()
}
