% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.toList}
\alias{.toList}
\title{Convert a string separated by commas into a vector of strings}
\usage{
.toList(a_string)
}
\arguments{
\item{a_string}{character representing a comma-separated list}
}
\value{
list of character
}
\description{
Attempts to parse items separated by double quotes
}
\examples{
toList('“Isolation, sample, environment”, isolation')
}
