% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-download.R
\name{download}
\alias{download}
\title{Download Something from BacDive}
\usage{
download(URL, userpwd = paste(get_credentials(), collapse = ":"))
}
\arguments{
\item{URL}{represented by a correctly encoded character string with spaces,
thanks to utils::URLencode.}

\item{userpwd}{A character string of the format
\code{BacDive_email:BacDive_password}. Retrieved from \code{.Renviron} my default,
but also used with something else by the tests.}
}
\value{
A serialised JSON string.
}
\description{
Download Something from BacDive
}
