% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Renviron_path}
\alias{get_Renviron_path}
\title{Construct Path of User's .Renviron File}
\usage{
get_Renviron_path()
}
\value{
A character vector representing the path to the user's \code{.Renviron}
file.
}
\description{
Construct Path of User's .Renviron File
}
