% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataById.R
\name{getDataByBacDiveId}
\alias{getDataByBacDiveId}
\title{Get BacDive Data by a BacDive ID}
\usage{
getDataByBacDiveId(
  access_object,
  id,
  template_path = "inst/extdata/template.csv",
  verbose = FALSE
)
}
\arguments{
\item{access_object}{BacDive Access Object BacDive access object}

\item{id}{integer BacDive ID}

\item{template_path}{character Path to the record template.
Defaults to inst/extdata/template.csv.}

\item{verbose}{character if TRUE write messages to console. Defaults to
FALSE.}
}
\value{
list of lists of BacDive Data
}
\description{
Get BacDive Data by a BacDive ID
}
\examples{
access_object <- authenticate('username', 'password')
getDataByBacDiveId(access_object, 12345)
}
