% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_searchType.R
\name{guess_searchType}
\alias{guess_searchType}
\title{Guess Search Type}
\usage{
guess_searchType(searchTerm, searchType)
}
\arguments{
\item{searchTerm}{Mandatory character string (in case of \code{sequence},
\code{culturecollectionno} or \code{taxon} \code{searchType})  or integer (in case of
\code{bacdive_id}), specifying what shall be searched for.}

\item{searchType}{Mandatory character string that specifies which type of
search will be performed (technically, which API endpoint). Can be
\code{bacdive_id} (default), \code{sequence}, \code{culturecollectionno} or \code{taxon}.}
}
\value{
a character vector containing the searchTerm again, and the possibly
second-guessed searchType
}
\description{
Guess Search Type
}
\examples{
guess_searchType(searchTerm = "DSM 319", searchType = "bacdive_id")
}
