% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prepare_Renviron.R
\name{prepare_Renviron}
\alias{prepare_Renviron}
\title{Prepare .Renviron File}
\usage{
prepare_Renviron()
}
\value{
Doesn't return any value, but writes (to and if necessary, creates)
the .Renviron file so users can type in their \code{BacDive_email} and
\code{BacDive_password} manually.
}
\description{
Prepare .Renviron File
}
\examples{
prepare_Renviron()
}
