% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve data from BacDive}
\usage{
retrieve_data(searchTerm, searchType = "bacdive_id", force = FALSE)
}
\arguments{
\item{searchTerm}{Mandatory character string (in case of \code{sequence},
\code{culturecollectionno} or \code{taxon} \code{searchType})  or integer (in case of
\code{bacdive_id}), specifying what shall be searched for.}

\item{searchType}{Mandatory character string that specifies which type of
search will be performed (technically, which API endpoint). Can be
\code{bacdive_id} (default), \code{sequence}, \code{culturecollectionno} or \code{taxon}.}

\item{force}{Logical. Whether or not the searchType should be enforced
strictly, even if it appears to mismatch the searchTerm. Please note:
forcing an apparently mismatched searchType will most likely result in an
error: \code{retrieve_data(searchTerm = "AJ000733", searchType = "bacdive_id", force = TRUE)} without specifying \code{searchType = "sequence"} should lead to
an internal re-specification, and execution of the intended search.}
}
\value{
EITHER (from an unambiguous searchTerm) a list of lists containing a
single BacDive datase,
OR (from a _am_biguous searchTerm) a numeric vector of BacDive-IDs, which
can be fed back into \code{retrieve_data()} to retrieve the individual data
sets.
}
\description{
Retrieve data from BacDive
}
\examples{
retrieve_data(searchTerm = 717)
  retrieve_data("AJ000733", searchType = "sequence")
  retrieve_data("DSM 319", "culturecollectionno")
  retrieve_data("Bacillus subtilis subtilis", searchType = "taxon")
}
