% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_search_results.R
\name{retrieve_search_results}
\alias{retrieve_search_results}
\title{Retrieve Results of an Advanced Search/Query}
\usage{
retrieve_search_results(queryURL)
}
\arguments{
\item{queryURL}{URL of a query prepared on \href{https://bacdive.dsmz.de/advsearch}{BacDive.DSMZ.de/AdvSearch},
as explained in the vignette "The
semi-automatic approach: retrieving data through a pre-configured advanced
search".}
}
\value{
A list of lists containing either a single BacDive dataset in case
the \code{searchTerm} was unambiguous (\code{bacdive_id}, \code{sequence},
\code{culturecollectionno}), or a large list containing all datasets that match
an ambiguous \code{searchTerm} (most \code{taxon}s).
}
\description{
Retrieve Results of an Advanced Search/Query
}
\examples{
  \donttest{data_miller <- retrieve_search_results(queryURL = "https://bacdive.dsmz.de/advsearch?site=advsearch&searchparams[78][contenttype]=text&searchparams[78][typecontent]=contains&searchparams[78][searchterm]=Miller&advsearch=search")}

}
