% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_input.R
\name{sanitise_input}
\alias{sanitise_input}
\title{Split taxon searchTerm}
\usage{
sanitise_input(searchTerm, searchType)
}
\arguments{
\item{searchTerm}{Mandatory character string (in case of \code{sequence},
\code{culturecollectionno} or \code{taxon} \code{searchType})  or integer (in case of
\code{bacdive_id}), specifying what shall be searched for.}

\item{searchType}{Mandatory character string that specifies which type of
search will be performed (technically, which API endpoint). Can be
\code{bacdive_id} (default), \code{sequence}, \code{culturecollectionno} or \code{taxon}.}
}
\value{
A character string with possibly sanitised searchType, see example.
}
\description{
Split taxon searchTerm
}
