% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{aggregate_result_IDs}
\alias{aggregate_result_IDs}
\title{Aggregate BacDive-IDs from a Paged List of Retrieved URLs}
\usage{
aggregate_result_IDs(results)
}
\arguments{
\item{results}{A list of paginated URLs returned by an ambiguous
\code{searchTerm} in \code{retrieve_data()}}
}
\value{
An integer vector of all BacDive IDs within the results.
}
\description{
Aggregate BacDive-IDs from a Paged List of Retrieved URLs
}
