% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_url.R
\name{construct_url}
\alias{construct_url}
\title{Construct URLs of API searches}
\usage{
construct_url(searchTerm, searchType = "bacdive_id", force = FALSE)
}
\arguments{
\item{searchTerm}{Mandatory character string (in case of \code{sequence},
\code{culturecollectionno} or \code{taxon} \code{searchType})  or integer (in case of
\code{bacdive_id}), specifying what shall be searched for.}

\item{searchType}{Mandatory character string that specifies which type of
search will be performed (technically, which API endpoint). Can be
\code{bacdive_id} (default), \code{sequence}, \code{culturecollectionno} or \code{taxon}.}

\item{force}{Logical. Whether or not the searchType should be enforced
strictly, even if it appears to mismatch the searchTerm. See examples.}
}
\description{
Construct URLs of API searches
}
