% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.formatRecord}
\alias{.formatRecord}
\title{Use template CSV to create functions to apply on BacDive records}
\usage{
.formatRecord(bacdive_data, record_template, debug = FALSE)
}
\arguments{
\item{bacdive_data}{list of lists}

\item{record_template}{data.frame of instructions for each row of data}

\item{debug}{logical print debug messages. Default to FALSE.}
}
\value{
list of data
}
\description{
Use template CSV to create functions to apply on BacDive records
}
\examples{
formatRecord(bacdive_data, template)
}
