% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getSequenceId}
\alias{.getSequenceId}
\title{Get sequence id and append to string separated by semicolons}
\usage{
.getSequenceId(bacdive_data, elements, database, verbose = FALSE)
}
\arguments{
\item{bacdive_data}{list of a BacDive data}

\item{elements}{character representing the elements column of the template}

\item{database}{character representing the sequence databases in
.getSequenceDatabases().}

\item{verbose}{logical print messages. Default to FALSE.}
}
\value{
vector of character
}
\description{
Get sequence id and append to string separated by semicolons
}
\examples{
.getSequenceId(bacdive_data, "Sequence information, 16S sequences", "ena")
}
