% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-download.R
\name{download}
\alias{download}
\title{Download Something from BacDive}
\usage{
download(URL)
}
\arguments{
\item{URL}{A correctly encoded character string, either from construct_url()
or a JSON list}
}
\value{
The R object unserialised from the downloaded JSON
}
\description{
Download Something from BacDive
}
\keyword{internal}
