% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-download.R
\name{download}
\alias{download}
\title{Download Something from BacDive}
\usage{
download(URL, user = get_credentials()[1],
  password = get_credentials()[2])
}
\arguments{
\item{URL}{A correctly encoded character string, either from construct_url()
or a JSON list}
}
\value{
The R object deserialised from the downloaded JSON
}
\description{
Both \code{user} and \code{password} are read from the .Renviron file by default.
}
\keyword{internal}
