% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Read BacDive Login Credentials from .Renviron}
\usage{
get_credentials()
}
\value{
A vector or the two character strings \code{BacDive_email}
and \code{BacDive_password} representing those two environment variables.
}
\description{
Read BacDive Login Credentials from .Renviron
}
\keyword{internal}
