% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-repair_escaping.R
\name{repair_escaping}
\alias{repair_escaping}
\title{Repair Singly Escaped Space Characters}
\usage{
repair_escaping(JSON, char)
}
\arguments{
\item{JSON}{A character string containing JSON with invalidly, singly
escaped space characters}
}
\value{
A character string with JSON and correctly, doubly escaped
space characters
}
\description{
Prevent "lexical error: invalid character inside string.", see
https://github.com/jeroen/jsonlite/issues/47 and
https://github.com/TIBHannover/BacDiveR/issues/43
}
