% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve (a) Dataset(s) from BacDive}
\usage{
retrieve_data(searchTerm, searchType = "taxon")
}
\arguments{
\item{searchTerm}{Mandatory character string (in case of \code{searchType = }
\code{sequence}, \code{culturecollectionno} or \code{taxon}) or integer (in case of
\code{bacdive_id}), specifying what shall be searched for.}

\item{searchType}{Mandatory character string that specifies which type of
search will be performed (technically, which Web Service endpoint). Can be \code{taxon}
(default), \code{bacdive_id}, \code{sequence}, or \code{culturecollectionno}.}
}
\value{
A list of lists containing either a single BacDive dataset in case
the \code{searchTerm} was unambiguous (\code{bacdive_id}, \code{sequence},
\code{culturecollectionno}), or a large list containing all datasets that match
an ambiguous \code{searchTerm} (most \code{taxon}s).
}
\description{
Retrieve (a) Dataset(s) from BacDive
}
\examples{
\donttest{
dataset_717 <- retrieve_data(searchTerm = 717, searchType = "bacdive_id")
}
\donttest{
dataset_DSM_319 <- retrieve_data(searchTerm = "DSM 319", searchType = "culturecollectionno")
}
\donttest{
dataset_AJ000733 <- retrieve_data(searchTerm = "AJ000733", searchType = "sequence")
}
\donttest{
datasets_Bh <- retrieve_data(searchTerm = "Bacillus halotolerans")
}
}
